requirements.txt
pywebview==3.8
pyinstaller==5.13.0
